backend/requirements.txt
# ───────────────────────────────────────────────────
# AI Wellness Buddy — Backend requirements
# ───────────────────────────────────────────────────

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Async DB
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0           # SQLite async driver (dev/testing)
asyncpg>=0.29.0             # PostgreSQL async driver (production)
alembic>=1.13.1             # Database migrations

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
passlib[bcrypt]>=1.7.4

# Config
pydantic>=2.7.0
pydantic-settings>=2.2.1
pydantic[email]>=2.7.0

# HTTP client (tests / frontend proxy)
httpx>=0.27.0

# Testing
pytest>=8.1.1
pytest-asyncio>=0.23.6
pytest-mock>=3.14.0

# ─── AI / ML stack (shared with root project) ───────
textblob>=0.17.1
transformers>=4.40.0
torch>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.4
